# Reactor-Plugin/openapi.yaml
openapi: 3.0.1
info:
  title: Code Generation Plugin
  description: A plugin that allows the user to generate code using ChatGPT.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /generate-code:
    post:
      operationId: generateCode
      summary: Generate code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/generateCodeRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generateCodeResponse'
  /Reactor/{username}:
    get:
      operationId: getInitialCode
      summary: Get the initial code for a specific user
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the user.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getCodeResponse'
  /current-code/{username}:
    get:
      operationId: getCurrentCode
      summary: Get the current code for a specific user
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the user.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getCodeResponse'

components:
  schemas:
    generateCodeRequest:
      type: object
      required:
      - prompt
      properties:
        prompt:
          type: string
          description: The prompt for the code generation.
          required: true
    generateCodeResponse:
      type: object
      properties:
        code:
          type: string
          description: The generated code.
          required: true
    getCodeResponse:
      type: object
      properties:
        code:
          type: string
          description: The returned code.